requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
openai-whisper==20231117
uvicorn==0.24.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
torchaudio
torchvision
modelscope
packaging
addict
datasets
simplejson
sortedcontainers
librosa
pydub
groq
python-dotenv
